build(phpcs): remove the blanket phpcs:ignoreFile suppressions and fix what they hid
Five files began with // phpcs:ignoreFile, which switches off all seventeen of
Plugin Check's PHPCS-backed checks - including all five security checks - on
the settings screen, both admin_post handlers, the diagnostic logger, the log
reader and the log admin page. A clean automated gate on those files was
meaningless because the tool was never allowed to look at them.

The five lines are gone and the findings they hid are fixed rather than
re-suppressed:

- The three grouped settings arrays are read through one helper that runs
  map_deep( ..., 'sanitize_text_field' ) over the value before anything looks
  at it (WordPress.Security.ValidatedSanitizedInput.InputNotSanitized).
- unlink() becomes wp_delete_file()
  (WordPress.WP.AlternativeFunctions.unlink_unlink).
- Six printf arguments in the log table are cast at the output site
  (WordPress.Security.EscapeOutput.OutputNotEscaped). All six were already
  integers reaching %d placeholders, so this is a gate fix, not a security fix,
  and it must not be described as one.
- The WP_DEBUG-gated error_log() of the CSS inventory is removed rather than
  annotated; the plugin has a proper opt-in diagnostic log for this.
- class-data-migration.php keeps its ABSPATH guard within the first 50 lines
  and no longer assembles an IN() list by interpolation, so its queries are
  fully prepared with no interpolated fragment at all.

Where a sniff genuinely misfires the suppression is now per line, names the
sniff, and carries a one-line justification, following the pattern the codebase
already uses at class-batch-convert-wizard.php:943.

Measured with the PHPCS bundled in Plugin Check 2.0.0 over the five files:
plugin-check.ruleset.xml and plugin-review.xml both report 0 errors and 0
warnings with annotations honoured (from 6 errors / 18 warnings before). With
--ignore-annotations the residue is 3 errors and 16 warnings, every one of them
covered by a named, justified per-line annotation. `wp plugin check` on the
built zip reports nothing in these files except the six MissingTranslatorsComment
errors, which are finding N2 and belong to a separate change.

+++ src/admin/class-conversion-log-admin.php
@@ -1,6 +1,4 @@
 <?php
-// phpcs:ignoreFile
-
 /**
  * Conversion Log admin UI page.
  *
@@ -67,8 +65,8 @@ class Conversion_Log_Admin {
 	 *
 	 * @param string $hook Current admin page hook suffix.
 	 */
-	public function enqueue_assets( string $hook ): void { // phpcs:ignore Generic.CodeAnalysis.UnusedFunctionParameter.Found
-		if ( empty( $_GET['page'] ) || self::MENU_SLUG !== sanitize_key( wp_unslash( $_GET['page'] ) ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+	public function enqueue_assets( string $hook ): void { // phpcs:ignore Generic.CodeAnalysis.UnusedFunctionParameter.Found -- Signature is fixed by the admin_enqueue_scripts hook; the screen is identified from the page query var instead.
+		if ( empty( $_GET['page'] ) || self::MENU_SLUG !== sanitize_key( wp_unslash( $_GET['page'] ) ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- Read-only screen check on an enqueue hook; nothing is written, so a nonce would be meaningless here.
 			return;
 		}
 
@@ -202,9 +200,9 @@ class Conversion_Log_Admin {
 		// Newest first for the table.
 		$log = array_reverse( $log );
 
-		$filter     = isset( $_GET['status'] ) ? sanitize_key( (string) $_GET['status'] ) : 'all'; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		$filter     = isset( $_GET['status'] ) ? sanitize_key( (string) $_GET['status'] ) : 'all'; // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- Display-only table filter behind a manage_options check; it changes no state.
 		$logging_on = Admin_Settings::is_logging_enabled();
-		$cleared    = isset( $_GET['blockshift_cleared'] ) && '1' === $_GET['blockshift_cleared']; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		$cleared    = isset( $_GET['blockshift_cleared'] ) && '1' === $_GET['blockshift_cleared']; // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- Post-redirect flag that only decides whether a notice is shown; the clear action itself is nonce-checked in handle_clear_log().
 
 		// Summary counts (computed before filter is applied).
 		$total_count = count( $log );
@@ -251,12 +249,12 @@ class Conversion_Log_Admin {
 
 		// JSONL diagnostic log state.
 		$jsonl_log_path  = Diagnostic_Logger::log_path();
-		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_operations_file_exists
+		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_operations_file_exists -- Read-only probe of the plugin's own log path while rendering a screen; initialising WP_Filesystem here would prompt for credentials on non-direct hosts.
 		$jsonl_log_exists = file_exists( $jsonl_log_path );
 		$jsonl_log_size   = $jsonl_log_exists ? (int) filesize( $jsonl_log_path ) : 0;
 		$jsonl_log_lines  = array();
 		if ( $jsonl_log_exists && $jsonl_log_size > 0 ) {
-			// phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents
+			// phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents -- Reading a local file the plugin wrote itself, not a remote resource; WP_Filesystem has no line-wise reader and would prompt for credentials.
 			$_all_lines      = file( $jsonl_log_path, FILE_IGNORE_NEW_LINES );
 			$_all_lines      = is_array( $_all_lines ) ? $_all_lines : array();
 			$jsonl_log_lines = array_slice( $_all_lines, -self::JSONL_LOG_TAIL );
@@ -436,7 +434,7 @@ class Conversion_Log_Admin {
 									<td style="min-width:140px;">
 										<?php if ( $total_w > 0 ) : ?>
 											<div class="pgs-progress pgs-progress--sm"><div class="pgs-progress__track"><div class="pgs-progress__fill<?php echo $is_partial ? ' pgs-progress__fill--warning' : ''; ?>" style="width:<?php echo esc_attr( $pct ); ?>%;"></div></div></div>
-											<div class="pgs-table__count"><?php printf( esc_html__( '%1$d / %2$d', 'migrate-off-elementor' ), $converted_w, $total_w ); ?></div>
+											<div class="pgs-table__count"><?php printf( esc_html__( '%1$d / %2$d', 'migrate-off-elementor' ), (int) $converted_w, (int) $total_w ); ?></div>
 										<?php else : ?>
 											<span class="pgs-table__muted">—</span>
 										<?php endif; ?>
@@ -448,7 +446,7 @@ class Conversion_Log_Admin {
 													<i data-icon="alert-triangle"></i>
 													<?php
 													$nt = count( $unsp_types );
-													printf( esc_html( _n( '%d unsupported type', '%d unsupported types', $nt, 'migrate-off-elementor' ) ), $nt );
+													printf( esc_html( _n( '%d unsupported type', '%d unsupported types', $nt, 'migrate-off-elementor' ) ), (int) $nt );
 													?>
 												</summary>
 												<ul style="margin:6px 0 0 16px;padding:0;font-size:var(--text-xs);color:var(--text-muted);list-style:disc;">
@@ -459,12 +457,12 @@ class Conversion_Log_Admin {
 											</details>
 											<?php if ( $empty_w > 0 ) : ?>
 												<div class="pgs-table__meta">
-													<?php printf( esc_html( _n( '+%d empty output', '+%d empty outputs', $empty_w, 'migrate-off-elementor' ) ), $empty_w ); ?>
+													<?php printf( esc_html( _n( '+%d empty output', '+%d empty outputs', $empty_w, 'migrate-off-elementor' ) ), (int) $empty_w ); ?>
 												</div>
 											<?php endif; ?>
 										<?php elseif ( $empty_w > 0 ) : ?>
 											<span class="pgs-issue">
-												<i data-icon="info"></i> <?php printf( esc_html( _n( '%d empty output', '%d empty outputs', $empty_w, 'migrate-off-elementor' ) ), $empty_w ); ?>
+												<i data-icon="info"></i> <?php printf( esc_html( _n( '%d empty output', '%d empty outputs', $empty_w, 'migrate-off-elementor' ) ), (int) $empty_w ); ?>
 											</span>
 										<?php else : ?>
 											<span class="pgs-issue pgs-issue--ok"><i data-icon="check"></i> <?php esc_html_e( 'None', 'migrate-off-elementor' ); ?></span>
@@ -480,8 +478,8 @@ class Conversion_Log_Admin {
 							<?php
 							printf(
 								esc_html__( 'Showing %1$d entries. Log retains the last %2$d entries; oldest are discarded automatically.', 'migrate-off-elementor' ),
-								count( $log ),
-								self::MAX_ENTRIES
+								(int) count( $log ),
+								(int) self::MAX_ENTRIES
 							);
 							?>
 						</div>
@@ -511,7 +509,7 @@ class Conversion_Log_Admin {
 								<?php
 								if ( $jsonl_log_exists && $jsonl_log_size > 0 ) {
 									echo esc_html( size_format( $jsonl_log_size ) );
-									// phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_operations_file_exists
+									// phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_operations_file_exists -- Read-only probe of the plugin's own rotated log while rendering a screen.
 									if ( file_exists( $jsonl_log_path . '.1' ) ) {
 										echo ' &nbsp;';
 										esc_html_e( '(rotated backup: .jsonl.1 also present)', 'migrate-off-elementor' );
